+++ Shell/Commands/InitCommand.cs
@@ -6,17 +6,19 @@ namespace Shell.Commands
     public class InitCommand : ICommand
     {
         public string version { get; set; }
-        public InitCommand(string Version)
+        public Logger logger { get; set; }
+        public InitCommand(string Version, Logger Logger)
         {
             version = Version;
+            logger = Logger;
         }
         public void Run(string[] arguments)
         {
-            Logger.Log("Starting Init Command", this.GetType().Name, LogType.INFO);
+            logger.Log("Starting Init Command", this.GetType().Name, LogType.INFO);
             var nameArgument = "main";
             if (arguments.Length == 2) { nameArgument = arguments[1];}
 
-            Logger.Log($"Creating {nameArgument} file", this.GetType().Name, LogType.INFO);
+            logger.Log($"Creating {nameArgument} file", this.GetType().Name, LogType.INFO);
             var fileName = nameArgument.Replace(".pirate", "");
             var file = File.CreateText($"./{fileName}.pirate");
             file.Write(String.Join(
@@ -27,7 +29,7 @@ namespace Shell.Commands
                 "}"
             ));
             file.Close();
-            Logger.Log($"Created {nameArgument} file", this.GetType().Name, LogType.INFO);
+            logger.Log($"Created {nameArgument} file", this.GetType().Name, LogType.INFO);
             Console.WriteLine($"\nCreated {fileName}.pirate");
         }