+++ Shell/Commands/RunCommand.cs
@@ -7,13 +7,15 @@ namespace Shell.Commands
     public class RunCommand : ICommand
     {
         public string version { get; set; }
-        public RunCommand(string Version)
+        public Logger logger { get; set; }
+        public RunCommand(string Version, Logger Logger)
         {
             version = Version;
+            logger = Logger;
         }
         public void Run(string[] arguments)
         {
-            Logger.Log("Starting Run Command", this.GetType().Name, LogType.INFO);
+            logger.Log("Starting Run Command", this.GetType().Name, LogType.INFO);
             var fileArgument = "main";
             if (arguments.Length >= 2) { fileArgument = arguments[1]; }
             var fileName = fileArgument.Replace(".pirate", "");
@@ -21,20 +23,20 @@ namespace Shell.Commands
 
             if (!exists)
             {
-                Logger.Log($"File \"{fileArgument}\" not provided or does not exist.", this.GetType().Name, LogType.ERROR);
+                logger.Log($"File \"{fileArgument}\" not provided or does not exist.", this.GetType().Name, LogType.ERROR);
                 Error($"File \"{fileArgument}\" not provided or does not exist.");
                 return;
             }
 
-            Logger.Log("Starting build", this.GetType().Name, LogType.INFO);
-            var buildCommand = new BuildCommand(version);
+            logger.Log("Starting build", this.GetType().Name, LogType.INFO);
+            var buildCommand = new BuildCommand(version, logger);
             buildCommand.Run(arguments);
-            Logger.Log("Completed Build", this.GetType().Name, LogType.INFO);
+            logger.Log("Completed Build", this.GetType().Name, LogType.INFO);
 
             var location = $"bin/pirate{version}";
-            Logger.Log($"Executing {fileName}.py", this.GetType().Name, LogType.INFO);
+            logger.Log($"Executing {fileName}.py", this.GetType().Name, LogType.INFO);
             var pythonEngine = new PythonEngine($"{location}/{fileName}.py");
-            var result = pythonEngine.InvokeMain("main"); //Possibly null reference;
+            var result = pythonEngine.InvokeMain("main", logger); //Possibly null reference;
         }
 
         public void Help()